Bloqueando Exclusão de Especialidade de Médico Quando Agendada

+++ src/main/java/com/mballem/curso/security/web/controller/MedicoController.java
@@ -68,8 +68,13 @@ public class MedicoController {
 	@GetMapping({"/id/{idMed}/excluir/especializacao/{idEsp}"})
 	public String excluirEspecialidadePorMedico(@PathVariable("idMed") Long idMed, 
 									   @PathVariable("idEsp") Long idEsp, RedirectAttributes attr) {
-		service.excluirEspecialidadePorMedico(idMed, idEsp);
-		attr.addFlashAttribute("sucesso", "Especialidade removida com sucesso.");
+		if(service.existeEspecialidadeAgendada(idMed, idEsp)) {
+			attr.addFlashAttribute("falha", "Tem consultas agendadas, exclusão negada.");
+		} else {
+			service.excluirEspecialidadePorMedico(idMed, idEsp);
+			attr.addFlashAttribute("sucesso", "Especialidade removida com sucesso.");
+		}
+		
 		
 		return "redirect:/medicos/dados";
 	}
